+++ ReleaseNotes.md
@@ -153,3 +153,4 @@ Please change your configuration according to [App.xaml](https://github.com/spie
 * New controls combining a `TextBox` and a `FileSystemDialog`
   * `TextBoxOpenDirectory`
 * New `MaterialWindow.TitleTemplate` property to customize the window title bar
+* New control `NavigationRail`